% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{calc_covariate_R2s}
\alias{calc_covariate_R2s}
\title{Calculate R2 of level 1, level 2, and block.}
\usage{
calc_covariate_R2s(data, ICCs = NULL, pooled = FALSE)
}
\arguments{
\item{data}{in canonical form.}

\item{ICCs}{List of ICC values, level 2 and level 3 in order.}

\item{pooled}{Calculate pooled Tx and Co groups, attempting to
remove systematic shift of average treatment effect.  If FALSE,
subset to control observations only.}
}
\description{
block is always the fixed effects for block.  No covariates
allowed at that level.  In all models block fixed effects are
included, so level 2 covarites R2 are additional explanatory power
beyond their representing block differences (e.g., they are
considered block centered).
}
