% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{calc_ICCs}
\alias{calc_ICCs}
\title{Calculate ICCs}
\usage{
calc_ICCs(data)
}
\description{
Use a multilevel model to calculate ICCs.  Treated units included,
using a constant treatment effect model.
}
