% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_weight_esimators.R
\name{canonical_weight_estimators}
\alias{canonical_weight_estimators}
\title{Estimate ATE via canonical weights on the cluster means}
\usage{
canonical_weight_estimators(formula, data = NULL, data_agg = NULL)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\value{
tibble of estimates using different varieties of the
  methods described in the paper.
}
\description{
These estimators use the cluster means and the table of weights
shown in the main paper to estimate the ATE across all combinations
of weights at the block and the cluster level.
}
\details{
There are no standard errors for this method.  It is purely to
validate how the other estimators can be represented as given with
the weight formula in the paper.
}
