% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_data_integrity}
\alias{check_data_integrity}
\title{Check data integrity}
\usage{
check_data_integrity(formula = NULL, data)
}
\description{
Ensure data has proper treatment variable, clusterIDs are defined,
and so forth.
}
