% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{deconstruct_control_formula}
\alias{deconstruct_control_formula}
\title{Get variables listed in control_formula formula.}
\usage{
deconstruct_control_formula(control_formula, data)
}
\value{
list of variable names
}
\description{
Get variables listed in control_formula formula.
}
