% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{deconstruct_var_formula}
\alias{deconstruct_var_formula}
\title{Process formula to get variable names}
\usage{
deconstruct_var_formula(formula, data)
}
\description{
Process formula to get variable names
}
