% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{expand_control_variables}
\alias{expand_control_variables}
\title{Take data with possibly categorical control variables and expand to
dummy variables.}
\usage{
expand_control_variables(data, control_formula)
}
\value{
List of three objects: the revised data frame, revised
  control_formula, and final list of names of control variables.
}
\description{
Take data with possibly categorical control variables and expand to
dummy variables.
}
