% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{get_estimand}
\alias{get_estimand}
\title{Get the estimand for a given method.}
\usage{
get_estimand(method, simple = TRUE)
}
\arguments{
\item{method}{Character vector of method name(s), as they appear
in `method_characteristics()`'s `method` column.}

\item{simple}{If TRUE, collapse "Cluster-Block"/"Person-Block"
weight labels down to "Cluster"/"Person".  If FALSE, collapse
only "Cluster-Cluster"/"Person-Person" (leaving "-Block" labels
distinct).}
}
\description{
Get the estimand for a given method.
}
\seealso{
method_characteristics()
}
