% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{has_singleton_clusters_agg}
\alias{has_singleton_clusters_agg}
\title{Any treatment arms with singleton cluster (tx or co)?}
\usage{
has_singleton_clusters_agg(datagg)
}
\arguments{
\item{datagg}{Aggregated data with blockID and Z columns.}
}
\value{
TRUE or FALSE
}
\description{
Used to avoid calculating nonsensical standard errors.
}
