% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{identify_singleton_blocks}
\alias{identify_singleton_blocks}
\title{Identify all tx and all co blocks}
\usage{
identify_singleton_blocks(formula, data)
}
\arguments{
\item{formula}{Notation for Y ~ Z | clusterID | blockID.  If NULL,
`data` is assumed to already be in canonical form.}

\item{data}{Dataframe to check.}
}
\value{
Dataframe of block IDs corresponding to all tx or all co
  blocks along with number of units and tx status.  Returns NULL if
  nothing.
}
\description{
Identify blocks that are all tx or all co.  Rows with missing
treatment indicators are dropped.  Rows with missing block
identifiers are considered all unique blocks and thus put into the
table of results.
}
