% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_models.R
\name{interacted_aggregated_estimators}
\alias{interacted_aggregated_estimators}
\title{Interacted linear regression models}
\usage{
interacted_aggregated_estimators(formula, data = NULL, control_formula = NULL)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\value{
Dataframe of the different weighting versions of this
  estimator (by number of schools (clusters) or simple average)
}
\description{
These linear models have siteID by treatment interaction terms.
The final ATE estimates are then weighted average of the block
(site) specific ATE estimates.
}
\details{
SEs come from the overall variance-covariance matrix.
}
