% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_nested}
\alias{is_nested}
\title{Check if one grouping variable nested in another}
\usage{
is_nested(clusterID, siteID)
}
\arguments{
\item{clusterID}{List of clusterIDs}

\item{siteID}{List of siteIDs}
}
\value{
TRUE if nested, FALSE otherwise.  If siteID NULL, always
  return TRUE.
}
\description{
Checks if clusterID nested in siteID, so each clusterID value
appears in only one siteID value.
}
