% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_regression_formula}
\alias{make_regression_formula}
\title{Make a canonical regression formula (no FE), possibly with control
variables.}
\usage{
make_regression_formula(
  Yobs = "Yobs",
  Z = "Z",
  clusterID = "clusterID",
  siteID = "siteID",
  control_formula = NULL,
  interacted = FALSE,
  FE = FALSE,
  cluster_RE = FALSE,
  data = NULL
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{Name of treatment variable (assumed to exist in data)}

\item{clusterID}{Name of cluster ID variable (assumed to exist in
data)}

\item{siteID}{Name of site ID variable (assumed to exist in data,
if this is not null).}

\item{control_formula}{What variables to control for, in the form
of "~ X1 + X2".}

\item{interacted}{TRUE means include treatment by site
interactions.  Will override FE flag and set FE to true.}

\item{FE}{TRUE means include site dummy variables.}

\item{cluster_RE}{Add a random effect term for clusterID}

\item{data}{Dataframe holding all variables to be used in formula.}
}
\value{
Something like "Yobs ~ 1 + Z" or "Yobs ~ 1 + Z + X1 + X2"
}
\description{
Make a canonical regression formula (no FE), possibly with control
variables.
}
