% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{method_characteristics}
\alias{method_characteristics}
\title{Get table of characteristics of all the methods implemented in this
package.}
\usage{
method_characteristics(include_weight = TRUE)
}
\arguments{
\item{include_weight}{If TRUE (default), include the `weight`
column (the estimand each method targets).  If FALSE, drop it.}
}
\value{
Dataframe with columns for method, weight, population,
  biased, and disfavored (whether we do not like the estimator due
  to odd weighting or instability).
}
\description{
Get table of characteristics of all the methods implemented in this
package.
}
\examples{
method_characteristics()

}
