% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDStdCRT_estimator.R
\name{MRStdCRT_estimator}
\alias{MRStdCRT_estimator}
\title{Estimate ATEs for a cluster RCT using the MRStdCRT pacakge}
\usage{
MRStdCRT_estimator(
  formula,
  data = NULL,
  control_formula = NULL,
  weight = c("Person", "Cluster")
)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\description{
NOTE: You will have to manually install this package via GitHub:
Use:
}
\details{
devtools::install_github("deckardt98/MRStdCRT")
}
