% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{print.clusterRCTstats}
\alias{print.clusterRCTstats}
\title{Pretty print result from describe cluster RCT call}
\usage{
\method{print}{clusterRCTstats}(x, ...)
}
\arguments{
\item{x}{A clusterRCTstats object.}

\item{...}{No extra options passed.}
}
\description{
Pretty print result from describe cluster RCT call
}
\seealso{
Other clusterRCTstats: 
\code{\link{as.data.frame.clusterRCTstats}()},
\code{\link{is.clusterRCTstats}()}
}
\concept{clusterRCTstats}
