% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_based_estimators.R
\name{schochet_overall_estimator}
\alias{schochet_overall_estimator}
\title{Given block specific estimates, calculate the ATE, associated
standard error, and design-based degrees of freedom.}
\usage{
schochet_overall_estimator(block_estimates)
}
\arguments{
\item{block_estimates}{Result of the
schochet_variance_formula_block() function call.}
}
\value{
List of ATE, SE, and degrees of freedom.
}
\description{
Given block specific estimates, calculate the ATE, associated
standard error, and design-based degrees of freedom.
}
