% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerandomize_simulation_code.R
\name{summarize_simulation_results}
\alias{summarize_simulation_results}
\title{Summarize set of compare_methods results across estimators}
\usage{
summarize_simulation_results(rps, summarize_results = "method")
}
\arguments{
\item{rps}{A list of data frames, each containing the results of a
compare_methods call.}

\item{summarize_results}{If "cross", summarize the results by
calculating range statistics across the set, grouping estimators
by estimand within each runID.  If "cross-agg", aggregate the
sets of ranges after doing this and return summary of that.  If
"method", summarize the results for each method. Otherwise just
stack the initial set of results (if not already stacked)}
}
\value{
A data frame containing the summarized results.
}
\description{
Summarize set of compare_methods results across estimators
}
