% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate data to the cluster level.}
\usage{
aggregate_data(data, control_formula = NULL)
}
\arguments{
\item{data}{Dataframe with 'clusterID' and 'Z' as columns.
'blockID' optional column.  This is data in the "canonical form".}
}
\value{
tibble of cluster-aggregated data, including Ybar, n,
  blockID, clusterID, and Z
}
\description{
Also implements checks to ensure that all clusters are entirely
treated or not treated.
}
\details{
If covariates specified, categorical covariates will be converted
to dummy variables and then averaged (even if they are only level-2
or level-3 covariates)
}
