% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_estimators.R
\name{aggregation_estimators}
\alias{aggregation_estimators}
\title{Estimate ATEs for cluster RCT using data aggregation.}
\usage{
aggregation_estimators(
  formula,
  data = NULL,
  control_formula = NULL,
  control_interacted = FALSE,
  aggregated = FALSE
)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\description{
This method aggregates data at the cluster level, and analyzes
using linear regression. Can optionally pass pre-aggregated data if
desired.
}
