% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{as.data.frame.clusterRCTstats}
\alias{as.data.frame.clusterRCTstats}
\alias{clusterRCTstats}
\title{Cast cluster RCT info result to data.frame}
\usage{
\method{as.data.frame}{clusterRCTstats}(x)
}
\arguments{
\item{x}{the clusterRCTstats object to covert}
}
\description{
Cast cluster RCT info result to data.frame
}
\seealso{
Other clusterRCTstats: 
\code{\link{is.clusterRCTstats}()},
\code{\link{print.clusterRCTstats}()}
}
\concept{clusterRCTstats}
