% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{calc_covariate_R2s}
\alias{calc_covariate_R2s}
\title{Calculate R2 of level 1, level 2, and district.}
\usage{
calc_covariate_R2s(data, pooled = FALSE)
}
\arguments{
\item{data}{in canonical form.}

\item{pooled}{Calculate pooled Tx and Co groups, attempting to
remove systematic shift of average treatment effect.  If FALSE,
subset to control observations only.}
}
\description{
District is always the fixed effects for district.  No covariates
allowed at that level.  In all models district fixed effects are
included, so level 2 covarites R2 are additional explanatory power
beyond their representing district differences (e.g., they are
considered district centered).
}
