% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare_methods}
\alias{compare_methods}
\title{Compare different estimators for a (possibly blocked) cluster
randomized trial.}
\usage{
compare_methods(
  formula,
  data = NULL,
  control_formula = NULL,
  weight = c("individual", "cluster"),
  include_MLM = TRUE,
  include_DB = TRUE,
  include_LM = TRUE,
  include_agg = TRUE,
  warn_missing = TRUE,
  include_dumb = FALSE,
  include_disfavored = FALSE,
  patch_data = TRUE,
  handle_singleton_blocks = c("drop", "pool", "fail"),
  include_method_characteristics = TRUE
)
}
\arguments{
\item{data}{frame holding Y, Z, B and (possibly a column with name
specified by blockID).}

\item{control_formula}{What variables to control for, in the form
of "~ X1 + X2".}

\item{include_MLM}{Include MLM estimators}

\item{include_DB}{Include Design-Based estimators (taken from
RCTYes documentation and prior literature).}

\item{include_LM}{Include Linear Model-based estimators (including
Huber-White SEs, etc.)}

\item{include_agg}{Include estimators applied to aggregated data,
aggregated at the cluster level.}

\item{include_dumb}{Include "dumb" estimators (i.e., those
interacted estimators that weight by both person and cluster or
vice versa).}

\item{patch_data}{If TRUE impute all missing covariates with mean
imputation (adding dummy variables as needed) via the
`patch_data()` method.  Will drop all rows with missing outcome,
treatment, or clustering info.  If FALSE do not do this.}

\item{include_method_characteristics}{Include details of the
methods (target estimands and sampling framework assumed) in the
return value.}

\item{Yobs}{vector observed outcomes (or column name in data)}

\item{Z}{vector of assignment indicators (1==treated) (or column
name in data)}

\item{B}{vector of block ids (or column name in data)}

\item{blockID}{block ids (variable name as string if data frame
passed) (if randomization blocks are nested in block).}
}
\value{
Dataframe of point estimates and standard errors for each
  method considered. If \code{include_method_characteristics=TRUE}
  also add some features of the methods as additional columns.
}
\description{
This function calculates the point estimates and SE estimates for a
variety of the estimators implemented by this package.
}
\examples{
data( fakeCRT )
compare_methods( Yobs ~ T.x | S.id | D.id, data=fakeCRT )

}
