% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{count_covariates}
\alias{count_covariates}
\title{Calculate number of level 1 and level 2 variables}
\usage{
count_covariates(
  data,
  control_formula,
  cluster_formula = NULL,
  just_count = FALSE
)
}
\arguments{
\item{data}{Data frame with outcome, treatment, clusterID, blockID,
and covariates.}

\item{control_formula}{Formula with covariates only (no outcome or
treatment).}

\item{cluster_formula}{Formula for cluster ID, of form ~ CID or ~
CID | BID.  In the latter case, will make interaction to ensure
all clusters have unique ID.  If null, assumes a clusterID and
blockID column exist in the data.}

\item{just_count}{If TRUE, return just counts of level 1 and level
2 variables.}
}
\value{
If just_count is FALSE (default), a tibble with columns:
  \item{variable}{Variable name.} \item{level}{Either "Level-1" or
  "Level-2".} If just_count is TRUE, a named vector with counts of
  level 1 and level 2 variables.
}
\description{
Given a control formula, determine which variables are level 1 and
which are level 2.  Level 1 variables vary within clusters, level 2
do not. If formula and data are provided, will first convert to
canonical form. If just_count is TRUE, will return a named vector
with counts of level 1 and level 2 variables.
}
