% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{describe_clusterRCT}
\alias{describe_clusterRCT}
\title{Describe a cluster RCT dataset}
\usage{
describe_clusterRCT(formula = NULL, data = NULL, control_formula = NULL)
}
\arguments{
\item{formula}{}
}
\description{
Calculate various statistics describing total clusters, etc.
}
