% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_based_estimators.R
\name{design_based_estimators}
\alias{design_based_estimators}
\title{Estimate ATE via design-based approach}
\usage{
design_based_estimators(
  formula,
  data = NULL,
  control_formula = NULL,
  weight = c("Person", "Cluster"),
  aggregated = FALSE,
  include_block_estimates = FALSE
)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}

\item{aggregated}{TRUE means data is already aggregated (and in
canonical form).  FALSE means it is not.}
}
\value{
tibble of estimates using different varieties of the
  methods described in the paper.
}
\description{
This follows "Design-Based Ratio Estimators and Central Limit
Theorems for Clustered, Blocked RCTs" by Peter Z. Schochet, Nicole
E. Pashley, Luke W. Miratrix, and Tim Kautz.
}
\details{
It runs a weighted linear regression, as discussed in that paper.
The paper discusses two possabilities: running on the aggregate, or
running on the individual.

In this implementation, we run the regression on the
cluster-aggregated data.
}
