% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_based_estimators.R
\name{design_based_estimators_individual}
\alias{design_based_estimators_individual}
\title{Estimate ATE via design-based approach}
\usage{
design_based_estimators_individual(
  formula,
  data = NULL,
  control_formula = NULL,
  weight = c("Person", "Cluster"),
  include_block_estimates = FALSE
)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\value{
tibble of estimates using different varieties of the
  methods described in the paper.
}
\description{
This function follows `design_based_estimators`, but runs the
regression on the individual level data.
}
