% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_docs.R
\docType{data}
\name{fakeCRT2}
\alias{fakeCRT2}
\title{Fake Cluster Randomized Trial data (set 2)}
\format{
A data frame with 1500 rows and 7 variables: \describe{
 \item{\code{S.id}}{integer School ID.}
 \item{\code{D.id}}{integer District ID.}
 \item{\code{Y1}}{double Potential outcome under treatment.}
 \item{\code{Y0}}{double Potential outcome under no treatment.}
 \item{\code{Yobs}}{double Observed outcome (think test score).)}
 \item{\code{T.x}}{integer Treatment assignment (1 treated, 0 control).} }
}
\usage{
fakeCRT2
}
\description{
This dataset is a three-level dataset with students in
 schools in districts, with schools randomized to treatment and
 control.  It is used to illustrate and test the clusterRCT pacakge.
}
\details{
These data were generated via the PUMP package.
}
\seealso{
\code{\link{fakeCRT}}
}
\keyword{datasets}
