% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_estimators.R
\name{gee_estimators}
\alias{gee_estimators}
\title{Estimate ATEs for a cluster RCT using a GEE approach}
\usage{
gee_estimators(
  formula,
  data = NULL,
  control_formula = NULL,
  weight = c("Person", "Cluster")
)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\description{
Estimate ATEs for a cluster RCT using a GEE approach
}
