% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{get_estimand}
\alias{get_estimand}
\title{Get the estimand for a given method.}
\usage{
get_estimand(method, simple = TRUE)
}
\description{
Get the estimand for a given method.
}
\seealso{
method_characteristics()
}
