% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{get_structure}
\alias{get_structure}
\title{Given data, return a data frame with the structure of the data}
\usage{
get_structure(formula = NULL, data)
}
\arguments{
\item{data}{frame holding Y, Z, B and (possibly a column with name
specified by blockID).}
}
\value{
Dataframe listing blockID, clusterID, treatment status, and
  number of units in the cluster.
}
\description{
Given data, return a data frame with the structure of the data
}
