% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_models.R
\name{interacted_aggregated_estimators}
\alias{interacted_aggregated_estimators}
\title{Interacted linear regression models}
\usage{
interacted_aggregated_estimators(formula, data = NULL, control_formula = NULL)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\value{
Dataframe of the different versions of this estimator (person and
  site weighted)
}
\description{
These linear models have siteID by treatment interaction terms.  The final ATE
estimates are then weighted average of the block (site) specific ATE
estimates.
}
\details{
#' If siteID passed, it will weight the RA blocks within site and then average
these site estimates.

SEs come from the overall variance-covariance matrix.
}
