% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_models.R
\name{interacted_linear_model_estimators}
\alias{interacted_linear_model_estimators}
\title{Estimate ATEs for a cluster RCT using linear model}
\usage{
interacted_linear_model_estimators(
  formula,
  data = NULL,
  control_formula = NULL,
  use_full_vcov = FALSE
)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\description{
Estimate district-average effects using an interacted FE model and
then aggregate.
}
