% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_estimators.R
\name{interacted_linear_model_estimators}
\alias{interacted_linear_model_estimators}
\title{Estimate ATEs for a cluster RCT using linear model}
\usage{
interacted_linear_model_estimators(
  formula,
  data = NULL,
  control_formula = NULL,
  weight = c("Person", "Cluster"),
  use_full_vcov = FALSE
)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}

\item{use_full_vcov}{TRUE/FALSE. When calculating standard errors,
should the full variance-covariance matrix of the block-level
estimates be used, or just the diagonal of standard errors?}
}
\description{
Estimate district-average effects using an interacted FE model and
then aggregate.
}
