% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{is.clusterRCTstats}
\alias{is.clusterRCTstats}
\alias{clusterRCTstats}
\title{Is object a clusterRCTstats object?}
\usage{
is.clusterRCTstats(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Is object a clusterRCTstats object?
}
\seealso{
Other clusterRCTstats: 
\code{\link{as.data.frame.clusterRCTstats}()},
\code{\link{print.clusterRCTstats}()}
}
\concept{clusterRCTstats}
