% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_nested}
\alias{is_nested}
\title{Check if one grouping variable nested in another}
\usage{
is_nested(clusterID, blockID)
}
\arguments{
\item{clusterID}{List of clusterIDs}

\item{blockID}{List of blockIDs}
}
\value{
TRUE if nested, FALSE otherwise.  If blockID NULL, always
  return TRUE.
}
\description{
Checks if clusterID nested in blockID, so each clusterID value
appears in only one blockID value.
}
