% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_estimators.R
\name{linear_model_estimators}
\alias{linear_model_estimators}
\title{Estimate ATEs for a cluster RCT using linear model}
\usage{
linear_model_estimators(
  formula,
  data = NULL,
  control_formula = NULL,
  weight = c("Person", "Cluster"),
  blockCRVE = FALSE
)
}
\arguments{
\item{formula}{Formula for outcome and treatment and nesting.  If
NULL, data is assumed to be in canonical form (see vignette for
further discussion).}
}
\description{
Estimate ATEs for a cluster RCT using linear model
}
