% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{make_block_table}
\alias{make_block_table}
\title{Make table of statistics by block}
\usage{
make_block_table(
  formula = NULL,
  data = NULL,
  control_formula = NULL,
  check_data_integrity = FALSE
)
}
\arguments{
\item{check_data_integrity}{TRUE means runs some checks and give
errors if data fails them (e.g., incorrectly processed treatment
vector.). FALSE means calculate statistics without these checks.}
}
\description{
Given individual/school/block data of a blocked, cluster RCT,
calculate statistics for each block (block).
}
