% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_clusterRCT.R
\name{make_site_table}
\alias{make_site_table}
\title{Make table of statistics by district}
\usage{
make_site_table(formula = NULL, data = NULL, control_formula = NULL)
}
\description{
Given individual/school/district data of a blocked, cluster RCT, calculate statistics for each block (district).
}
