% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_data_set.R
\name{patch_data_set}
\alias{patch_data_set}
\title{Fill in missing values and convert to canonical form}
\usage{
patch_data_set(formula = NULL, data = NULL, control_formula = NULL)
}
\description{
This method will fill in means for all missing covariate values and
return a dataset where rows with missing outcome, cluster or block
membership, and treatment assignment are all dropped and everything
else is filled in.  To do this, it will convert categorical
covariates to a set of dummy variables (dropping the reference
group).
}
\details{
It will drop missing data indicators that are colinear with prior
missing data indicators.
}
