% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerandomize_simulation_code.R
\name{rerandomize}
\alias{rerandomize}
\title{Return new treatment assignment based on passed one.}
\usage{
rerandomize(Z, clusterID, blockID = NULL)
}
\value{
vector of 1s and 0s.
}
\description{
Return new treatment assignment based on passed one.
}
