% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerandomize_simulation_code.R
\name{run_rerandomize_simulation}
\alias{run_rerandomize_simulation}
\title{Run a finite sample simulation by permuting treatment assignment
labels within each block.}
\usage{
run_rerandomize_simulation(
  formula,
  data,
  R = 100,
  summarize_results = TRUE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula object specifying the model to estimate.}

\item{data}{A data frame containing the data to analyze.}

\item{R}{The number of simulations to run.}

\item{summarize_results}{If TRUE, summarize the results of the
compare_methods call, FALSE return raw estimates.}

\item{parallel}{If TRUE, run simulations in parallel.}

\item{...}{Additional arguments to pass to compare_methods.}
}
\description{
Call compare_methods repeatidly on the resulting treatment-permuted
data.
}
